+++ tests/Reader/NativePhpReaderTest.php
@@ -18,7 +18,7 @@ use Ajgl\Csv\Reader\NativePhpReader;
 /**
  * @author Antonio J. García Lagar <aj@garcialagar.es>
  */
-class NativePhpReaderTest extends \PHPUnit_Framework_TestCase
+class NativePhpReaderTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @var NativePhpReader